requirements.txt
Flask>=1.1.1,<1.2.0
psycopg2>=2.8.4,<2.9.0
flask_migrate>=2.5.2,<2.6.0
python-dotenv>=0.10.3,<0.11.0
Flask-Cors>=3.0.8, <3.1.0
Flask-JWT-Extended>=3.24.1,<3.25.0
PyJWT>=1.7.1,<1.8.0
Flask-SQLAlchemy>=2.4.1,<2.5.0
Flask-RESTful>=0.3.7, <0.4.0
passlib>=1.7.2,<1.8.0
